When defining signs, check icon path for existence
sign define does not seem to check, that the given icon path exists when
a sign is defined, so check it manually. This possibly only happens when
the gui is started and is skipped when Vim startsup.

Add an additional check, that the give path exists and if not, define
the sign without the icon.